Format all files.
VSCode started prettifying files with Prettier, so this commit
introduces a new NPM command `npm run fmt` which runs prettier against
all .JS, .TS and .TSX files in the repository. This should be broadly
compatible with the way VSCode was formatting (using prettier). In the
future it may pay to introduce a config file for this to be more sane
about what we're formatting and why.